QuoteAutomation/pyproject.toml
[project]
name = "recommendationenginetest"
version = "0.1.0"
description = "RecommendationEngine using crewAI"
authors = [{ name = "Vladimir", email = "vladimir@elementsagents.com" }]
requires-python = ">=3.10,<=3.12"
dependencies = [
    "crewai-tools>=0.32.1",
    "crewai[tools]>=0.86.0,<1.0.0",
    "pinecone-client[grpc]>=2.2.4",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0"
]

[project.scripts]
recommendationenginetest = "recommendationenginetest.main:run"
run_crew = "recommendationenginetest.main:run"
train = "recommendationenginetest.main:train"
replay = "recommendationenginetest.main:replay"
test = "recommendationenginetest.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recommendationenginetest"]